:recycle: split cli tool worker isolation helper
